add a prediction script
This adds a `predict.py` script to do inference with the model,
predicting sql from questions without attempting to evaluate
accuracy (which would require that the questions be labeled).

The `train.py` script is tweaked slightly so that the methods
within it can be used by `predict.py`.

The `annotate_ws.py` script is also tweaked slightly so that
it can be run on user data.

Two utilities, `add_csv.py` and `add_question.py` are added
to give a quick way to set up some new tables and questions.

To minimize changes to existing code, I add unlabeled questions
with a dummy label, as empty as I could make it while still
having working code.